STY: move tests for listgen
